OvmfPkg: Implement TdxProbLib instance in OvmfPkg
For Td guest the initial mode on ResetVector is 32bit protected mode.
While for Non-Td guest the initial mode is 16bit real mode. Based on
this difference a TD flag is written to the Mailbox[0x10] in ResetVector.
So that the TdxProbeLib can probe this flag for Td or Non-Td guest.

This is a temporary solution to probe Td or Non-Td. When the related
TDX CPUID leaf is ready, this library will be refined to invoke CPUID
to probe the Td or Non-Td.