OvmfPkg: Build page table in run-time instead of using Rom page table
In standard OvmfPkg, page table is built in run-time. While in the
previous TDVF implementation, it uses the built-in page table. To be
compatible with the standard OvmfPkg, TDVF builds the page table in
run-time too.

Standard OvmfPkg supports 4-level page table, so it consumes 6 pages.
Td guest supports both 4-level and 5-level page table, so it consumes
7 pages. To meet this requirement, the page table in MEMFD is expanded
from 0x6000 to 0x7000.

Caution should be taken that page table can only be built in run-time
once, either by BSP or AP. So a flag (TDX_WORK_AREA_PGTBL_READY) is
imported to indicate whether the page table is ready. In Td guest BSP
is respoinsible for building the page table and then set the flag. APs
should wait for the flag then continue the work flow.

Some other code refine are made too.