OvmfPkg: Alter MADT in TdxDxe.
In order to maintain the simplicity of the AcpiPlatformDxe, the edit to
MADT read from Qemu need to be moved out.

A protocol notify is registered in the entrypoint of TdxDxe.
AcpiPlatformDxe will install a protocol instance while the ACPI tables
provided by Qemu is ready.

The registered function will traverse the ACPI table list to find the
original MADT from Qemu. After the new MADT is configured and installed,
the original one will be uninstalled.