OvmfPkg: Update the TDX page table region
Tdx supports both 4-level and 5-level paging table based on GPAW (48 or 52)
Its corresponding page table needs 6 page or 7 page. In previous
implementation PcdOvmfSecPageTablesSize is expanded to 7 page. But
it is not a good solution. The current solution is to use an extra page
to hold the top level Page Directory Pointers (1 * 256 TB entry)