MdePkg: Refine IoFifo for Tdx
In the previous IoLib IoFifo read/write will trigger #VE in Tdx guests.
It depends on the InitializeCpuExceptionHandlers. In some situation it
is not called before the IoFifo operation. QEMU fw_cfg operatioin in
PlatformPei is the example. So IoFifo in Tdx is refined to a serial call
of TdIoRead/TdIoWrite.