MdeModulePkg: Avoid using GCD entry of removed space
When removing unaccepted memory space entry start from 0x100000000, the
entry with merge with adjacent NonExist entry and the base address of
the entry will be changed, then accepting an illegal memory space will
cause a crash.

To fix this and prevent the potential failure of memory acceptance, remove
the unaccepted memory space until memory is accepted successfully, and do
not use the original entry pointer.

In addition, hard-code the memory space attribute of newly added system
memory with EFI_MEMORY_XP, EFI_MEMORY_RO, EFI_MEMORY_RP and
EFI_MEMORY_CPU_CRYPTO to align with the other system memory entries and be
able to merge with them.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>