OvmfPkg: Re-install the MADT table and append MPWK structure
During the boot process, MADT is downloaded and installed from QEMU
by virtual firmware. MPWK is the structure which describes the Mailbox
information to the OS, including the relocated address of the mailbox.
So MPWK should be created by virtual firmware and appended to the MADT.

In the previous implementation MADT table is re-constructed from scratch.
It is because TDVF assumes vCpu index is the Apic Id. But QEMU contructs
the MADT with the ApicId. This is correct if there is only one socket.
But if there are more sockets in the system, vCpu index cannot be the
ApicId.

So this commit is to clone the MATD table and append the newly created
MPWK structure after it.

Signed-off-by: Min Xu <min.m.xu@intel.com>