OvmfPkg/TdxStartupLib: Add SpeculationBarrier to mitigate bounds check bypass
Bounds check bypass uses speculative operations that occur while the
processor is checking whether an input is in bounds, such as checking
if the index of an array being read is within acceptable values. It
takes advantage of memory accesses to out of bound memory that are
performed speculatively before the bounds check resolves. These memory
accesses can be used in certain circumstances to leak information to
the attacker.

Software can insert a speculation stopping barrier between a bounds
check and a later operation that could cause a speculative side channel.
SpeculationBarrier() is such a barrier to mitigate the bounds check
bypass.

Signed-off-by: Min Xu <min.m.xu@intel.com>